+++ harvey/messages.py
@@ -7,6 +7,11 @@ from harvey.config import Config
 
 
 class Message:
+    # Emojis - used for Slack messages with sensible defaults if slack is not in use
+    work_emoji = ':hammer_and_wrench:' if Config.use_slack else ''
+    success_emoji = ':white_check_mark:' if Config.use_slack else 'Success!'
+    failure_emoji = ':skull_and_crossbones:' if Config.use_slack else 'Failure!'
+
     @staticmethod
     def send_slack_message(message: str):
         """Send a Slack message via a Slackbot."""